refactor: preparation for uniting Winners and Criteria
